Fix parquet writer crash on partitions with >127 distinct series
The series column was dictionary-encoded with an int8 index (max 127 distinct values). NSE's
historical bhavcopy carries well over 100 series codes; the 2023 partition reached 126, and the
next new code made every further write raise 'ArrowInvalid: dictionaries cannot be combined',
halting a 5-year backfill mid-2023. The 2025 partition only had 18 series, which is why the
earlier 15-month backfill never hit it.

- series index widened to int16
- upsert_partition/read_partition read with the FILE's schema then cast, so the many partitions
  already on disk (written with int8) stay extendable
- regression tests, including extending a genuinely int8-written partition

+++ backend/src/stk/store/parquet/schema.py
@@ -31,7 +31,11 @@ BARS_DAILY_SCHEMA = pa.schema(
         pa.field("symbol", pa.string(), nullable=False),
         pa.field("security_id", pa.int32(), nullable=True),
         pa.field("isin", pa.string(), nullable=True),
-        pa.field("series", pa.dictionary(pa.int8(), pa.string()), nullable=True),
+        # int16, NOT int8: an int8 index holds only 127 distinct values, and NSE's historical
+        # bhavcopy carries well over 100 series codes (bond/SME/T+0 series...). The 2023
+        # partition reached 126; the next new code made every further write fail with
+        # "dictionaries cannot be combined", halting a backfill mid-2023.
+        pa.field("series", pa.dictionary(pa.int16(), pa.string()), nullable=True),
         pa.field("instrument_type", pa.dictionary(pa.int8(), pa.string()), nullable=False),
         pa.field("open", pa.float64(), nullable=False),
         pa.field("high", pa.float64(), nullable=False),
